# examples/rest-service-chi/docs/openapi.yaml
# OpenAPI docs
openapi: 3.0.0
info:
  title: Example
  contact:
    name: Vitaly Lobchuk
    email: vitaly@lobchuk.ru
  version: v1.0.0
paths:
  /profiles:
    post:
      tags:
      - profile
      summary: Создать профиль
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                address:
                  description: |2+

                  type: string
                  example: abc
                firstName:
                  description: |2+

                  type: string
                  example: abc
                lastName:
                  description: |2+

                  type: string
                  example: abc
                zip:
                  description: |2+

                  type: integer
                  example: 1
          application/xml:
            schema:
              type: object
              properties:
                address:
                  description: |2+

                  type: string
                  example: abc
                firstName:
                  description: |2+

                  type: string
                  example: abc
                lastName:
                  description: |2+

                  type: string
                  example: abc
                zip:
                  description: |2+

                  type: integer
                  example: 1
              xml:
                name: profile
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  profile:
                    $ref: '#/components/schemas/Profile'
                xml: {}
            application/xml:
              schema:
                type: object
                properties:
                  profile:
                    $ref: '#/components/schemas/Profile'
                xml: {}
  /profiles/:id:
    delete:
      tags:
      - profile
      summary: Удалить профиль
      parameters:
      - in: path
        name: id
        description: |2+

        required: true
        schema:
          description: |2+

          type: string
          example: abc
      responses:
        200:
          description: OK
  /profiles/:id/file:
    get:
      tags:
      - profile
      summary: DownloadFile
      parameters:
      - in: path
        name: id
        description: |2+

        required: true
        schema:
          description: |2+

          type: string
          example: abc
      responses:
        200:
          description: OK
components:
  schemas:
    Address:
      properties:
        apt:
          description: |
            Apartment
          type: integer
          example: 1
        street:
          description: |
            Street
          type: string
          example: abc
    ErrorDefaultError:
      type: object
      properties:
        errorText:
          type: string
          example: abc
    ErrorServerError:
      type: object
      properties:
        code:
          type: string
          example: abc
        errorText:
          type: string
          example: abc
    Profile:
      properties:
        address:
          $ref: '#/components/schemas/Address'
        fistName:
          description: |
            Surname
          type: string
          example: abc
        lastName:
          description: |
            Name
          type: string
          example: abc